- #36306, crc32() function result differs on 32-bit and 64-bit platforms

+++ ext/standard/crc32.c
@@ -38,7 +38,7 @@ PHP_NAMED_FUNCTION(php_if_crc32)
 	for (len += nr; nr--; ++p) {
 	    CRC32(crc, *p);
 	}
-	RETVAL_LONG(~crc);
+	RETVAL_LONG(~ (long) crc);
 }
 /* }}} */
